6.0

+++ Frameworks/MediaPlayer.framework/MPInlineAudioTransportControls.h
@@ -6,7 +6,7 @@
     int _style;
 }
 
-- (id)buttonImageForPart:(unsigned int)arg1;
+- (id)buttonImageForPart:(unsigned long long)arg1;
 - (id)hitTest:(struct CGPoint { float x1; float x2; })arg1 withEvent:(id)arg2;
 - (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1 style:(int)arg2;
 - (void)layoutSubviews;
