+++ Frameworks/MediaPlayer.framework/MPPadFullScreenTransportControls.h
@@ -6,14 +6,14 @@
 }
 
 - (void)_itemTypeAvailableNotification:(id)arg1;
-- (id)buttonImageForPart:(unsigned int)arg1;
+- (id)buttonImageForPart:(unsigned long long)arg1;
 - (void)dealloc;
 - (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
 - (void)layoutSubviews;
 - (id)newVolumeSlider;
 - (void)reloadForAdditions:(id)arg1 removals:(id)arg2 animate:(BOOL)arg3;
-- (void)setDisabledParts:(unsigned int)arg1;
+- (void)setDisabledParts:(unsigned long long)arg1;
 - (void)setItem:(id)arg1;
-- (void)setVisibleParts:(unsigned int)arg1 animated:(BOOL)arg2;
+- (void)setVisibleParts:(unsigned long long)arg1 animated:(BOOL)arg2;
 
 @end
