+++ PrivateFrameworks/Message.framework/_MFSecCMSEncoder.h
@@ -5,13 +5,13 @@
 @class NSMutableData;
 
 @interface _MFSecCMSEncoder : MFBufferedDataConsumer {
-    int _SecCMSError;
+    long _SecCMSError;
     struct SecCmsEncoderStr { } *_encoder;
     struct SecCmsMessageStr { } *_message;
     NSMutableData *_singleShot;
 }
 
-@property(readonly) int lastSecCMSError;
+@property(readonly) long lastSecCMSError;
 
 - (void)_appendBytes:(const void*)arg1 length:(unsigned long)arg2;
 - (int)appendData:(id)arg1;
@@ -20,6 +20,6 @@
 - (void)done;
 - (id)initForEncryptionWithCompositionSpecification:(id)arg1 error:(id*)arg2;
 - (id)initForSigningWithSender:(id)arg1 compositionSpecification:(id)arg2 error:(id*)arg3;
-- (int)lastSecCMSError;
+- (long)lastSecCMSError;
 
 @end
